cloud-api/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
google-cloud-aiplatform>=1.38.1
pydub==0.25.1
numpy==1.24.3
requests==2.31.0
pytz==2024.1
gunicorn